refactor: extract internal/repl (Phase 2 step 7 of package reorg)
Move runREPL and all of its slash-command handlers (handleConnect,
handleDisconnect, handleKeys, handleSetCommand, printHelp, printContext,
printConfigInfo, reconnectMCPTransport, toggleOutputVerbose,
formatDuration), the live-feed auto-launcher (maybeLaunchLiveFeed,
waitForLiveFeedURL), and the /browserfeed VNC viewer
(cmd_browserfeed.go) into a new internal/repl package.

Public surface: a single repl.Run(ag, cliAgent, quietMode, version)
entrypoint. The Version constant is plumbed in as a parameter rather
than reaching back into package main.

Drop redundant prefixes in moved code:
  ShowBrowserFeed           → showBrowserFeed (now package-private)
  ShowBrowserFeedFromStream → showBrowserFeedFromStream
  printREPLHelp             → printHelp
  runREPL                   → Run

main.go shrinks from ~1960 to ~400 lines and now contains only
flag/subcommand routing, agent construction, and one-shot/positional
prompt handling.
