+++ internal/repl/browserfeed_test.go
@@ -1,4 +1,4 @@
-package main
+package repl
 
 import "testing"
