+++ internal/repl/live_feed_test.go
@@ -1,4 +1,4 @@
-package main
+package repl
 
 import (
 	"context"
